Project restructure, child libraries
Former-commit-id: 8d7ffeca1e5b72f11d68ec064c8ea86948a0072a

+++ bin/templates.sh
@@ -7,25 +7,43 @@ while [ -h "$SOURCE" ] ; do SOURCE="$(readlink "$SOURCE")"; done
 DIR="$( cd -P "$( dirname "$SOURCE" )/.." && pwd )"
 cd "$DIR"
 
+FORCE="$1"
+
 function tmpl {
+  echo "updating [$2] templates"
+  mv "$ftgt" "$fsrc"
+  if [ "$1" = "npntemplate" ]; then
+    cd npntemplate
+    rm -rf gen
+    hero -extensions .html,.sql -source "html" -pkgname $1 -dest "gen/npntemplate"
+    cd ..
+  else
+    rm -rf $3
+    hero -extensions .html,.sql -source "$2" -pkgname $1 -dest $3
+  fi
+}
+
+function check {
   fsrc="tmp/$1.hashcode"
   ftgt="tmp/$1.hashcode.tmp"
 
-  if [ ! -d "gen/$1" ]; then
+  if [ ! -d "$3" ]; then
     rm -f "$fsrc"
   fi
 
   find -s "$2" -type f -exec md5sum {} \; | md5sum > "$ftgt"
 
   if cmp -s "$fsrc" "$ftgt"; then
-    rm "$ftgt"
+    if [ "$FORCE" = "force" ]; then
+      tmpl $1 $2 $3 $4
+    else
+      rm "$ftgt"
+    fi
   else
-    echo "updating [$2] templates"
-    mv "$ftgt" "$fsrc"
-    rm -rf gen/$1
-    hero -extensions .html,.sql -source "$2" -pkgname $1 -dest gen/$1
+    tmpl $1 $2 $3 $4
   fi
 }
 
-tmpl "components" "web/components"
-tmpl "templates" "web/templates"
+check "npntemplate" "npntemplate/html" "npntemplate/gen"
+check "components" "web/components" "gen/components"
+check "templates" "web/templates" "gen/templates"
